requirements/development.txt
-r base.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
flake8==6.1.0
black==23.12.1
isort==5.13.2
pylint==3.0.3

# Development tools
ipython==8.18.1
ipdb==0.13.13
watchdog==3.0.0

# Documentation
sphinx==7.2.6

# Web framework & Database
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
sqlalchemy-cockroachdb==2.0.3

# Utilities
colorlog==6.8.0
coloredlogs==15.0.1
prettytable==3.10.2
tabulate==0.9.0
requests==2.31.0
pytz==2023.3
python-dotenv==1.0.0
pydantic==2.5.0
PyJWT==2.8.0
cryptography==42.0.7
pycryptodome==3.20.0
psycopg2-binary==2.9.9
sshkey-tools==0.11.3
sendgrid==6.11.0
